Add all additional methods from TreeTraversalMdAdditional to TreeTraversal. Fix incorrect ConsoleWriteLine. Use local functions. Create constructor for Node, which takes an id. Remove unnecessary curly braces. Remove redundant CreateTree method and just put the code into the constructor. Remove unnecessary comment. Remove the now redundant TreeTraversalMdAdditional. Sort the methods so that they have the same order as in the text.
